feat: add dispatch logging support to lock adapters
- Introduced a new DispatchLogging module for logging cron job dispatch attempts.
- Updated MySQLAdapter and PostgresAdapter to include the DispatchLogging module.
- Removed redundant log_dispatch_attempt and parse_lock_key methods from adapters, leveraging the new module instead.
- Enhanced observability of cron job dispatches for better auditing.
