+++ rails_cron/.reek.yml
@@ -46,11 +46,10 @@ detectors:
     exclude:
       - RailsCron::Registry#callable?
       - RailsCron::Lock::PostgresAdapter#calculate_lock_id
-      - RailsCron::Lock::PostgresAdapter#parse_lock_key
+      - RailsCron::Lock::DispatchLogging#parse_lock_key
       - RailsCron::Lock::PostgresAdapter#cast_to_boolean
       - RailsCron::Lock::RedisAdapter#generate_lock_value
       - RailsCron::Lock::MySQLAdapter#cast_to_boolean
-      - RailsCron::Lock::MySQLAdapter#parse_lock_key
       - RailsCron::Lock::MySQLAdapter#normalize_lock_name
   FeatureEnvy:
     enabled: true
