+++ rails_cron/lib/rails_cron/lock/postgres_adapter.rb
@@ -7,6 +7,7 @@
 
 require 'digest'
 require 'socket'
+require_relative 'dispatch_logging'
 
 module RailsCron
   module Lock
@@ -39,6 +40,8 @@ module RailsCron
     #     config.lock_adapter = RailsCron::Lock::PostgresAdapter.new(log_dispatch: true)
     #   end
     class PostgresAdapter < Adapter
+      include DispatchLogging
+
       attr_reader :log_dispatch
 
       ##
@@ -114,33 +117,6 @@ module RailsCron
         hash = Digest::MD5.digest(key).unpack1('Q>')
         # Convert to signed 64-bit integer
         hash > 9_223_372_036_854_775_807 ? hash - 18_446_744_073_709_551_616 : hash
-      end
-
-      def log_dispatch_attempt(key)
-        cron_key, fire_time = parse_lock_key(key)
-        node_id = Socket.gethostname
-
-        ::RailsCron::CronDispatch.create!(
-          key: cron_key,
-          fire_time: fire_time,
-          dispatched_at: Time.current,
-          node_id: node_id,
-          status: 'dispatched'
-        )
-      rescue StandardError => e
-        raise LockAdapterError, "Failed to log dispatch for #{key}: #{e.message}"
-      end
-
-      def parse_lock_key(key)
-        # Lock key format: "namespace:dispatch:cron_key:fire_time"
-        # Parse by splitting on colon: remove namespace, "dispatch", then rejoin remaining parts as key
-        parts = key.split(':')
-        fire_time_unix = parts.pop.to_i
-        2.times { parts.shift } # Remove namespace and "dispatch"
-        cron_key = parts.join(':')
-        fire_time = Time.at(fire_time_unix)
-
-        [cron_key, fire_time]
       end
     end
   end
